Refactor unit tests by removing duplicates and grouping regression tests
- Consolidated regression tests (`regr_*.rs`) into appropriate `semantic_*.rs` and `codegen_*.rs` files.
- Removed duplicated incomplete type and redefinition tests from `semantic_negative.rs` where they were already covered by `semantic_types.rs`, `semantic_scope.rs`, and `semantic_validation.rs`.
- Deleted `guardian_typedef_redefinition.rs` and moved tests into `semantic_scope.rs`.
- Grouped related error cases into consolidated `#[test]` functions to reduce test suite execution overhead (e.g. `test_typedef_redefinition_errors`, `test_linkage_mismatches`, `test_incomplete_type_prohibitions`).
- Validated with `cargo llvm-cov` that code coverage did not regress.
