+++ src/tests/semantic_usual_arithmetic_conversions.rs
@@ -46,3 +46,31 @@ fn test_usual_arithmetic_conversions() {
     let (_, result_long_long_unsigned_long) = run_pipeline(source_long_long_unsigned_long, CompilePhase::Mir);
     assert!(result_long_long_unsigned_long.is_ok());
 }
+
+#[test]
+fn test_mixed_sign_comparison_long_long() {
+    let source = r#"
+        #include <stdint.h>
+        int main() {
+            uint64_t u = 0xFFFFFFFFFFFFFFFBULL; // -5 if signed
+            int64_t s = 1LL;
+            if (u < s) return 1; // Fails if signed comparison
+            return 0; // Passes if unsigned comparison
+        }
+    "#;
+    assert_eq!(crate::tests::codegen_common::run_c_code_exit_status(source), 0);
+}
+
+#[test]
+fn test_mixed_sign_comparison_int_long() {
+    let source = r#"
+        #include <stdint.h>
+        int main() {
+            uint32_t u = 0xFFFFFFFFU;
+            int64_t s = 1LL;
+            if (u < s) return 1;
+            return 0;
+        }
+    "#;
+    assert_eq!(crate::tests::codegen_common::run_c_code_exit_status(source), 0);
+}
